Split FetchConversionRates job up into 2 jobs
